move FindCmpr component
