+++ camper_app/src/components/CamperCard.js
@@ -1,11 +1,12 @@
 import React from "react";
 import "./CamperCard.css";
 import { useState, useEffect } from "react";
-
 import { getAllCampers } from "../api/getAllCampers";
 import { getCampersByType } from "../api/getCampersByType";
 import { Loader } from "./Loader";
 import { NotificationManager } from "react-notifications";
+import { Card } from "./Card";
+import { FindCmpr } from "./FindCmpr";
 
 export const CamperCard = () => {
   const [campers, setCampers] = useState([]);
@@ -58,26 +59,3 @@ export const CamperCard = () => {
 
 
 
-function FindCmpr({ getCamperType }) {
-  function handleSelectType(e) {
-    e.preventDefault();
-    getCamperType(e.target.value);
-  }
-
-  return (
-    <div classNAme="find">
-      <h2 className="find__title">Znajdź campera</h2>
-      <p class="find_desc">Wybierz z listy: </p>
-      <div class="select">
-        <select id="cars" name="camperType" onChange={handleSelectType}>
-          <option value="allcapers">All Campers</option>
-          <option value="campervan">Campervan</option>
-          <option value="integra">Integra</option>
-          <option value="polintegra">Półintegra</option>
-          <option value="alkowa">Alkowa</option>
-        </select>
-        <div class="select_arrow"></div>
-      </div>
-    </div>
-  );
-}
